Update security-providers.md
Added the word Enter to:

Prior to applying a new or changing an existing security provider, make sure that every users are disconnected from the data source. If you are changing an existing Shared Passphrase or Certificate, please note that users will get back access to the data source when they "ENTER" the new Shared Passphrase or Certificate on their computer.